fix(Materials): switch to mobile particle shader
There is an issue with certain headsets like the quest using the
Oculus Integration CameraRig where the standard particle shader does
not render correctly, causing the rest of the screen to render black.

This just switches the default particle shader to the mobile shader.
It doesn't look as good, but its hardly noticable and will fix this
issue for anyone else and anyone who wants the better looking shader
can just provide custom materials.